♻️ Refactor SessionDetailPage file structure for better organization
- Move schema.ts from Messages to SessionDetailPage root level
- Relocate utility functions to SessionDetailPage/utils directory
- Create index.ts barrel file for cleaner imports
- Update all import paths to reflect new structure
- Improve component hierarchy and reduce deep nesting

This change makes shared utilities and schemas more accessible across
SessionDetailPage components while maintaining clear boundaries.
