+++ frontend/apps/app/components/SessionDetailPage/components/Chat/components/Messages/AiMessage/ToolCalls/ToolCall/ToolCall.tsx
@@ -10,9 +10,9 @@ import {
 } from '@liam-hq/ui'
 import { type FC, useMemo, useState } from 'react'
 import { MarkdownContent } from '../../../../../../../../MarkdownContent'
+import type { ToolCall as ToolCallType } from '../../../../../../../schema'
+import { extractResponseFromMessage } from '../../../../../../../utils'
 import markdownStyles from '../../../Markdown.module.css'
-import type { ToolCall as ToolCallType } from '../../../schema'
-import { extractResponseFromMessage } from '../../../utils/extractResponseFromMessage'
 import styles from './ToolCall.module.css'
 
 type Props = {
