+++ frontend/apps/app/components/SessionDetailPage/schema.ts
