+++ frontend/apps/app/components/SessionDetailPage/utils/extractReasoningFromMessage.ts
