+++ frontend/apps/app/components/SessionDetailPage/utils/extractResponseFromMessage.ts
