+++ frontend/apps/app/components/SessionDetailPage/utils/extractToolCallsFromMessage.ts
