+++ frontend/apps/app/components/SessionDetailPage/utils/getContentString.ts
